Fixes #8846 - Have the V2 api controller extend Api::V2::BaseController
Experiencing a bug where code in Foreman's Api::V2::BaseController is not being
inherited so before filters like check_content_type weren't being executed.
This change has the v2 api controller extend Api::V2::BaseController and as a
result, the Api::ApiController code was moved into a module so it could be
included.
